# Handle TCP disconnection.