--- a/tests/Serilog.Sinks.Datadog.Logs.Tests/Serilog.Sinks.Datadog.Logs.Tests.csproj
+++ b/tests/Serilog.Sinks.Datadog.Logs.Tests/Serilog.Sinks.Datadog.Logs.Tests.csproj
@@ -1,0 +1,15 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net5.0</TargetFramework>
+    <SignAssembly>true</SignAssembly>
+    <AssemblyOriginatorKeyFile>..\..\Serilog.Sinks.Datadog.Logs.snk</AssemblyOriginatorKeyFile>
+  </PropertyGroup>
+  <ItemGroup>
+    <PackageReference Include="NUnit" Version="3.13.2" />
+    <PackageReference Include="NUnit3TestAdapter" Version="4.0.0" />
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.3.0" />
+    <ProjectReference Include="../../src/Serilog.Sinks.Datadog.Logs/Serilog.Sinks.Datadog.Logs.csproj" />
+  </ItemGroup>
+
+</Project>