This PR handles TCP disconnection issues which lead to missing logs. See https://github.com/DataDog/serilog-sinks-datadog-logs/issues/21 for more details.
This PR is based from https://github.com/DataDog/serilog-sinks-datadog-logs/pull/20.

As an alternative solution, when the service disconnect, `TcpClient.Available` is greater than `0` when the service disconnect (at least in some cases). I did not implement this solution as I do not want to make assumption about what the server sends.
As `netstandard1.3` is deprecated  and mono on MacOS are not so common, it is better to use `GetActiveTcpConnections` to check if the tcp connection is established. 

Note: 
 - On Windows `GetActiveTcpConnections` return IPv4 adresses whereas `TcpClient.LocalEndPoint` and `TcpClient.RemoteEndPoint` return IPv6 adresses. 
 - The fix is not implemented for  `netstandard1.3` and for mono on MacOS.
 - There is a tiny time window between the call to `IsConnectionClosed()` and `_stream.Write(data, 0, data.Length);` where a disconnection to the TCP connection is not detected. As `_stream.Write(data, 0, data.Length);` does not throw when the TCP connection is not established, there is no way to detect this case.