@olivielpeau 2021-09-07
From our offline chat: let's add this new logic behind a config flag, turned off by default (long-term, all users should move to HTTP anyway). Worst-case, if this logic doesn't work well, it would constantly re-create even when not needed, which is not great.