@ogaca-dd 2021-09-07
> how costly is this logic (roughly)? It's called every time a batch of logs is submitted (so every 2 seconds or 50 logs according to the code), so in high throughput scenarios it would be called very often: at 1000 logs/sec, it would be called 20 times/second.

Serializing 50 logs requires more CPU than checking 100 TCP connections.

Small benchmark: sending 500 logs with log line length = 500.
Before:
```
| Method |     Mean |    Error |   StdDev |
|------- |---------:|---------:|---------:|
|  TCP | 963.4 ms | 41.73 ms | 121.7 ms | 924.7 ms |
```
After
```
| Method |     Mean |    Error |   StdDev |
|------- |---------:|---------:|---------:|
|  TCP | 971.4 ms | 47.95 ms | 140.6 ms |
```

> would it be difficult to add tests? For example ConnectionMatcher looks like logic that'd make sense to test and would be easy to test.

Done